Fix implicit property assignment error
Further examples of the implicit property assignment issue.

The problem appears when you have a campaign property set on the token
with the same name as the variable that you are assigning local to a
macro.

For instance, in the case of these we have base properties: "Damage",
"Wounds", etc. In our macros we would assign the result of a
user-defined function call to the (presumptive) local variable "damage"
or "wounds" respectively. The expected behavior is that this is locally
scoped to the macro, but in fact what appears to be happening is a
case-insensitive comparison occurs between "Damage" and "damage",
evaluates them to be the same, and then for some unknowable reason
thinks I want to assign the property on the token when I call
[h:damage= ... ].