Add roll helper overlay with combat modifiers
The roll helper overlay allows a user to bring up an overlay on top of
the screen that is draggable and collapsible which will allow them to
specify certain roll modifiers that should apply to their character.

Currently the roll helper has a single tab and it focuses on combat
modifiers.